Add files via upload
I conducted an internal IT security audit for Botium Toys as part of my portfolio for the Google Cybersecurity Professional Certificate, assessing the company’s infrastructure, risks, and compliance with regulations such as PCI DSS for online payments and GDPR for EU business operations. The work included reviewing the audit scope, goals, asset inventory, and risk assessment, and completing a controls and compliance checklist to identify potential vulnerabilities and mitigation strategies.